requirements.txt
Flask==3.0.2
gunicorn==21.2.0
groq==1.0.0
google-api-python-client==2.118.0
youtube-transcript-api==0.6.2
apify-client==1.6.2
python-dotenv==1.0.1
